[WEJBHTTP-140] Refactoring - Renaming TransactionConstants to Constants.
