+++ transaction/src/main/java/org/wildfly/httpclient/transaction/HttpSubordinateTransactionHandle.java
@@ -18,12 +18,12 @@
 
 package org.wildfly.httpclient.transaction;
 
+import static org.wildfly.httpclient.transaction.Constants.READ_ONLY;
 import static org.wildfly.httpclient.transaction.RequestType.XA_BEFORE_COMPLETION;
 import static org.wildfly.httpclient.transaction.RequestType.XA_COMMIT;
 import static org.wildfly.httpclient.transaction.RequestType.XA_FORGET;
 import static org.wildfly.httpclient.transaction.RequestType.XA_PREPARE;
 import static org.wildfly.httpclient.transaction.RequestType.XA_ROLLBACK;
-import static org.wildfly.httpclient.transaction.TransactionConstants.READ_ONLY;
 
 import io.undertow.client.ClientRequest;
 import io.undertow.client.ClientResponse;
