+++ transaction/src/main/java/org/wildfly/httpclient/transaction/RequestBuilder.java
@@ -23,17 +23,17 @@ import static io.undertow.util.Headers.CONTENT_TYPE;
 import static java.net.URLEncoder.encode;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.wildfly.httpclient.common.Protocol.VERSION_PATH;
+import static org.wildfly.httpclient.transaction.Constants.EXCEPTION;
+import static org.wildfly.httpclient.transaction.Constants.NEW_TRANSACTION;
+import static org.wildfly.httpclient.transaction.Constants.RECOVERY_FLAGS;
+import static org.wildfly.httpclient.transaction.Constants.RECOVERY_PARENT_NAME;
+import static org.wildfly.httpclient.transaction.Constants.TIMEOUT;
+import static org.wildfly.httpclient.transaction.Constants.TXN_CONTEXT;
+import static org.wildfly.httpclient.transaction.Constants.XID;
+import static org.wildfly.httpclient.transaction.Constants.XID_LIST;
 import static org.wildfly.httpclient.transaction.RequestType.UT_BEGIN;
 import static org.wildfly.httpclient.transaction.RequestType.XA_COMMIT;
 import static org.wildfly.httpclient.transaction.RequestType.XA_RECOVER;
-import static org.wildfly.httpclient.transaction.TransactionConstants.EXCEPTION;
-import static org.wildfly.httpclient.transaction.TransactionConstants.NEW_TRANSACTION;
-import static org.wildfly.httpclient.transaction.TransactionConstants.RECOVERY_FLAGS;
-import static org.wildfly.httpclient.transaction.TransactionConstants.RECOVERY_PARENT_NAME;
-import static org.wildfly.httpclient.transaction.TransactionConstants.TIMEOUT;
-import static org.wildfly.httpclient.transaction.TransactionConstants.TXN_CONTEXT;
-import static org.wildfly.httpclient.transaction.TransactionConstants.XID;
-import static org.wildfly.httpclient.transaction.TransactionConstants.XID_LIST;
 
 import io.undertow.client.ClientRequest;
 import io.undertow.util.HeaderMap;
